Rename Orgbook files to "Component"
