+++ app/package.json
@@ -32,7 +32,7 @@
     "components:clean": "cd ../components && npm run clean",
     "components:purge": "cd ../components && npm run purge",
     "frontend:ci": "cd frontend && npm ci",
-    "frontend:purge": "cd frontend && npm run purge",
+    "frontend:purge": "cd frontend && npm run purge && npm run purge:formio",
     "migrate": "npm run migrate:latest",
     "migrate:down": "knex migrate:down",
     "migrate:latest": "knex migrate:latest",
